Move MiniscreenButton to Miniscreen level
