+++ docs/api_pitop_device.rst
@@ -137,5 +137,5 @@ Here is an example demonstrating 2 ways to make use of these buttons:
 Class Reference: pi-top [4] Miniscreen Button
 ---------------------------------------------
 
-.. autoclass:: pitop.miniscreen.buttons.MiniscreenButton
+.. autoclass:: pitop.miniscreen.MiniscreenButton
     :inherited-members:
