+++ pitop/miniscreen/__init__.py
@@ -1,5 +1,4 @@
 from .miniscreen import Miniscreen
-from .buttons import MiniscreenButton
 
 # Deprecated:
 from .oled import OLED
