+++ pitop/miniscreen/buttons/__init__.py
@@ -1,5 +1,3 @@
-from .buttons import MiniscreenButton
-
 # Deprecated
 from .buttons import (
     Buttons,
