+++ pitop/miniscreen/buttons/buttons.py
@@ -1,3 +1,5 @@
+from pitop.miniscreen import MiniscreenButton
+
 from pitopcommon.ptdm import (
     PTDMSubscribeClient,
     Message
@@ -6,65 +8,6 @@ from pitopcommon.lock import PTLock
 from pitopcommon.singleton import Singleton
 
 import atexit
-
-
-class MiniscreenButton:
-    """
-    Represents one of the 4 buttons around the miniscreen display on a pi-top [4].
-    Should not be created directly - instead, use :class:`pitop.miniscreen.Miniscreen`.
-    """
-
-    def __init__(self):
-        # State parameters
-        self._is_pressed = False
-        # Event-based functions
-        self._when_pressed = None
-        self._when_released = None
-
-    @property
-    def is_pressed(self):
-        """
-        Get or set the button state as a boolean value.
-
-        :rtype: bool
-        """
-        return self._is_pressed
-
-    @is_pressed.setter
-    def is_pressed(self, value):
-        self._is_pressed = value
-
-    @property
-    def when_pressed(self):
-        """
-        Get or set the 'when pressed' button state callback function.
-        When set, this callback function will be invoked when this event happens.
-
-        :type callback: Function
-        :param callback:
-            Callback function to run when a button is pressed.
-        """
-        return self._when_pressed
-
-    @when_pressed.setter
-    def when_pressed(self, callback):
-        self._when_pressed = callback
-
-    @property
-    def when_released(self):
-        """
-        Get or set the 'when released' button state callback function.
-        When set, this callback function will be invoked when this event happens.
-
-        :type callback: Function
-        :param callback:
-            Callback function to run when a button is released.
-        """
-        return self._when_released
-
-    @when_released.setter
-    def when_released(self, callback):
-        self._when_released = callback
 
 
 class MiniscreenButtonLegacy(MiniscreenButton):
